% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{as_disagg}
\alias{as_disagg}
\title{Disaggregate from districts to zones.  Creates separate tables
for each subscriber class, purpose, and time of day.}
\usage{
as_disagg(asTable, equivTbl)
}
\arguments{
\item{asTable}{Table of AirSage data as provided in their
format.  Should have 9 fields:
\itemize{
\item Origin_Zone
\item Destination_Zone
\item Start_Date
\item End_Date
\item Aggregation
\item Subscriber_Class
\item Purpose
\item Time_of_Day
\item Count
}}

\item{equivTbl}{Equivalency table created by make_equivTbl()}
}
\description{
Disaggregate from districts to zones.  Creates separate tables
for each subscriber class, purpose, and time of day.
}

