% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_matrix.R
\name{balance_matrix}
\alias{balance_matrix}
\title{Balance matrix}
\usage{
balance_matrix(df, from = "origin", to = "destination")
}
\arguments{
\item{df}{A matrix in \code{data_frame} format with a column for origin and a
column for destination.}

\item{from}{A character string identifying the origin column, defaults to
\code{origin}}

\item{to}{A character string identifying the destination column, defaults
to \code{destination}}
}
\value{
A \code{data_frame} of the same format as \code{df}, with the matrix
  cores balanced.
}
\description{
This function takes an input matrix in origin/destination form and balances
it so that it is symmetric according to the following formula
\eqn{(A + A^T) / 2 }
}
\examples{
balance_matrix(test_oddf)

}
\author{
Kyle Ward
}
