% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{calc_perc}
\alias{calc_perc}
\title{Calculate the percent each centroid makes up of the district}
\usage{
calc_perc(equivLyr)
}
\arguments{
\item{equivLyr}{Equivalency Layer created by make_equivLyr()}
}
\description{
Calculate the percent each centroid makes up of the district
}

