% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids}
\alias{centroids}
\title{Sample Model Centroids (Hickory)}
\format{A SpatialPointsDataFrame}
\source{
Hickory TDM
}
\usage{
centroids
}
\description{
A layer containing the Hickory NC model centroid points and data.
}
\keyword{datasets}
