% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{explode}
\alias{explode}
\title{Explodes the district-to-district table into a zone-to-zone table}
\usage{
explode(asTable, equivLyr)
}
\arguments{
\item{asTable}{Table of AirSage data as provided in their standard
format.  Should have 9 fields:
\itemize{
\item Origin_Zone
\item Destination_Zone
\item Start_Date
\item End_Date
\item Aggregation
\item Subscriber_Class
\item Purpose
\item Time_of_Day
\item Count
}}

\item{equivLyr}{Equivalency layer updated by calc_perc()}
}
\description{
Explodes the district-to-district table into a zone-to-zone table
}

