% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{format}
\alias{format}
\title{Formats columns and values in the exploded table.}
\usage{
format(expTbl)
}
\arguments{
\item{expTbl}{Exploded, zonal table returned by explode}
}
\value{
Returns the final table.
}
\description{
Formats columns and values in the exploded table.
}

