% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{make_equivLyr}
\alias{make_equivLyr}
\title{Create Equivalency Layer}
\usage{
make_equivLyr(centroids, districts)
}
\arguments{
\item{centroids}{SpatialPointsDataFrame of model centroids.  Can
include the external stations (use the model node layer).
Must also contain the following fields:
\describe{
\item{ZONEID}{Unique identifier for each row/zone}
\item{EXTSTATION}{Externals stations marked with a 1}
\item{SE}{"Stuff" in the zone to use for weighted disaggregation. Can be
any combination of residential and/or employment info.
For example:
\itemize{
  \item SE = Households + Employment
  \item SE = Number of Parcels
  \item etc.
}}
\item{VOLUME}{For external stations, this column will contain a measure
of the volume/count at that station.}
}}

\item{districts}{SpatialPointsDataFrame of AirSage districts. Must be a polygon
layer.  Must contain the following field:
\describe{
  \item{DISTRICTID}{Unique identifier for each row/district}
}}
}
\value{
A SpatialPointsDataFrame of TAZ centroids with the District
   ID appended.
}
\description{
Create Equivalency Layer
}
