% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{make_equivTbl}
\alias{make_equivTbl}
\title{Create Equivalency Table}
\usage{
make_equivTbl(modelZones, tazID, asDistricts, distID)
}
\arguments{
\item{modelZones}{Path to a shapefile of model zones.  Must
include the external stations, too.  Use the node layer.
Must also contain the following fields:
\describe{
\item{zID}{Unique identifier for each row}
\item{Ext}{Externals stations marked with a 1}
\item{SE}{"Stuff" in the zone to use for disaggregation. Can be any
combination of residential and/or employment info.
For example:
\itemize{
\item SE = Households + Employment
\item SE = Number of Parcels
\item etc.
}
}
}}

\item{tazID}{TAZ ID Number}

\item{asDistricts}{Path to a shapefile of AirSage districts.
Must be a polygon layer.}

\item{distID}{The district ID that will be tagged to the zones}
}
\value{
A data frame with a row for each model zone specifying
   it's AirSage district
}
\description{
Create Equivalency Table
}

