% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airsage.R
\name{read_airsage}
\alias{read_airsage}
\title{Read an AirSage trip matrix}
\usage{
read_airsage(file)
}
\arguments{
\item{file}{Path to an AirSage CSV file}
}
\value{
A tibble with AirSage values coded in and tidy formats
}
\description{
Read an AirSage trip matrix
}
