% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_adjacent_trips.R
\name{remove_adjacent_trips}
\alias{remove_adjacent_trips}
\title{Remove adjacent external zone trips}
\usage{
remove_adjacent_trips(asTable, centroids = NULL, districts,
  ee_filter = NULL)
}
\arguments{
\item{asTable}{Table of AirSage data as provided in their standard
format.  Should have 9 fields:
\itemize{
\item Origin_Zone
\item Destination_Zone
\item Start_Date
\item End_Date
\item Aggregation
\item Subscriber_Class
\item Purpose
\item Time_of_Day
\item Count
}}

\item{centroids}{SpatialPointsDataFrame of model centroids.  Can
include the external stations (use the model node layer).
Must also contain the following fields:
\describe{
\item{ZONEID}{Unique identifier for each row/zone}
\item{EXTSTATION}{Externals stations marked with a 1}
\item{SE}{"Stuff" in the zone to use for weighted disaggregation. Can be
any combination of residential and/or employment info.
For example:
\itemize{
  \item SE = Households + Employment
  \item SE = Number of Parcels
  \item etc.
}}
\item{VOLUME}{For external stations, this column will contain a measure
of the volume/count at that station.}
}}

\item{districts}{SpatialPointsDataFrame of AirSage districts. Must be a polygon
layer.  Must contain the following field:
\describe{
  \item{DISTRICTID}{Unique identifier for each row/district}
}}

\item{ee_filter}{Dataframe with a `from` and `to` column. Removes flows from
included district pairs. Used to remove EE flows that do not enter the
model region.}
}
\value{
A `list` with four elements:
  - `corrected_flows` A `data_frame` of the same structure as `asTable`, but
  with adjacent and intra-district AirSage flows removed.
  - `original_external_trips` The original number of external zone trips.
  - `intrazonal_external_trips` The number of removed intrazonal trips in
  external districts.
  - `adjacent_external_trips` The number of adjacent-zone flows removed
  from the source data.
}
\description{
External trips to adjacent external zones should not be loaded onto the
region's model network.  Similarly, external trips made within the same
external zone are also removed.
}
\details{
If the `centroids` layer is not provided, then there must be a
  field in `districts$EXTSTATION` identifying external station zones with
  a value of 1 for external zones and 0 for internal zones.
}
