% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_external_od.R
\name{scale_external_od}
\alias{scale_external_od}
\title{Scale an external trips matrix}
\usage{
scale_external_od(od_matrix, counts)
}
\arguments{
\item{od_matrix}{Matrix in \code{data_frame} format, with columns \code{origin},
\code{destination}, and \code{volume}. See \code{\link{test_oddf}}}

\item{counts}{A \code{data_frame} with a \code{station} column indicating
which external station the count represents, and a \code{aawdt} column
containing the count volume. See \code{\link{test_counts}}}
}
\value{
A \code{data_frame} of the same structure as \code{od_matrix}, with columns
  added representing the scaled OD flows.
}
\description{
In modern big data applications, there is a high probability that the source
OD matrix (potentially including IE trips) will not match external station
counts. For example, StreetLight provides OD matrices that are normalized for
data disclosure and seasonality, and that therefore do not represent the actual
flows in the system. This function takes an input matrix and rescales it to
match counts at an external via an Iterative Proportional Fitting (IPF,
FRATAR, or raking) algorithm.
}
\details{
The function balances the input \code{od_matrix} and assumes that
  flows are symmetrical in a day.
}
