% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_external_stations.R
\name{split_external_stations}
\alias{split_external_stations}
\title{Split external zone trips between stations}
\usage{
split_external_stations(asTable, ext_stations)
}
\arguments{
\item{asTable}{An AirSage table from which adjacent zone external trips
have already been removed.}

\item{ext_stations}{A data_frame representing external stations with three
columns: 
  - `n` The centroid ID in the highway network
  - `zone` The AirSage zone the node is in
  - `awdt` The traffic count at the external stations}
}
\description{
The AirSage zones will typically have several external stations within
each zone. This function distributes these trips between the stations
according to the AWDT observed at each station. It does not rebalance
the flows to match the AWDT.
}
