% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{write}
\alias{write}
\title{Breaks up and writes out the exploded table}
\usage{
write(expTbl)
}
\arguments{
\item{expTbl}{Exploded, zonal table returned by explode}
}
\value{
expTbl Returns nothing, but writes out individual tables.
}
\description{
Breaks up and writes out the exploded table
}

