% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{as_disagg}
\alias{as_disagg}
\title{Disaggregate from districts to zones.  Creates separate tables
for each subscriber class, purpose, and time of day.}
\usage{
as_disagg(asTable, centroids, districts, tod_equiv)
}
\arguments{
\item{asTable}{Table of AirSage data as provided in their standard
format.  Should have 9 fields:
\itemize{
\item Origin_Zone
\item Destination_Zone
\item Start_Date
\item End_Date
\item Aggregation
\item Subscriber_Class
\item Purpose
\item Time_of_Day
\item Count
}}

\item{centroids}{SpatialPointsDataFrame of model centroids.  Must
include the external stations (use the model node layer).
Must also contain the following fields:
\describe{
\item{ZONEID}{Unique identifier for each row/zone}
\item{EXTSTATION}{Externals stations marked with a 1}
\item{SE}{"Stuff" in the zone to use for disaggregation. Can be any
combination of residential and/or employment info.
For example:
\itemize{
  \item SE = Households + Employment
  \item SE = Number of Parcels
  \item etc.
}}
\item{VOLUME}{For external stations, this column will contain a measure
of the volume/count at that station.}
}}

\item{districts}{SpatialPointsDataFrame of AirSage districts. Must be a polygon
layer.  Must contain the following field:
\describe{
  \item{DISTRICTID}{Unique identifier for each row/district}
}}

\item{tod_equiv}{Named list specifying the TOD used by the AirSage data.
For example:
\itemize{
  \item EA = "H00:H06"
  \item AM = "H06:H09"
  \item MD = "H09:H15"
  \item PM = "H15:H18"
  \item EV = "H18:H24"
}}
}
\value{
A tbl_df() object the contains the trips from origin zones to
   destination zones.  The functions also writes out tables to your working
   directory.
}
\description{
Disaggregate from districts to zones.  Creates separate tables
for each subscriber class, purpose, and time of day.
}

