% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{Sample AirSage Districts (Hickory)}
\format{A SpatialPolygonsDataFrame}
\source{
Hickory NC
}
\usage{
districts
}
\description{
A layer containing the Hickory NC AirSage district polygons.
}
\keyword{datasets}

