% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsage.R
\name{format}
\alias{format}
\title{Formats columns and values in the exploded table.}
\usage{
format(expTbl, tod_equiv)
}
\arguments{
\item{expTbl}{Exploded, zonal table returned by explode}

\item{tod_equiv}{Named list specifying the TOD used by the AirSage data.
For example:
\itemize{
  \item EA = "H00:H06"
  \item AM = "H06:H09"
  \item MD = "H09:H15"
  \item PM = "H15:H18"
  \item EV = "H18:H24"
}}
}
\value{
Returns the final table.
}
\description{
Formats columns and values in the exploded table.
}

